[NFCI][SYCL] Move `IsProperty[*Value]` to `property[_value].hpp`
This should make it easier to follow the logic and understand if template
parameters (like `PropertyT`) refer to keys or values.
