+++ sycl/include/sycl/ext/oneapi/properties/property.hpp
@@ -264,6 +264,13 @@ struct IsCompileTimeProperty
           std::is_base_of_v<property_key_base_tag, PropertyT> &&
           std::is_base_of_v<compile_time_property_key_base_tag, PropertyT>> {};
 
+// Checks if a type is either a runtime property or if it is a compile-time
+// property
+template <typename T> struct IsProperty {
+  static constexpr bool value =
+      IsRuntimeProperty<T>::value || IsCompileTimeProperty<T>::value;
+};
+
 // Trait for property compile-time meta names and values.
 template <typename PropertyT> struct PropertyMetaInfo {
   // Some properties don't have meaningful compile-time values.
